fix(views.py): Fixed bug that would not allow pre-screen xlsx export. Used linter to reformat views.py.
Fixed a bug where pre-screen exporting as xlsx would produce an error
where 'colspan' is not part of a method's arguments. Fixed another bug
that attempted to use 'strftime' method on a string which should be used
on a datatime object instead, changed to just use the string.
Reformatted the views.py file with a linter to be more readable while
bug hunting.